# packages/api-types/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/commerce/restaurants": {
      "get": {
        "operationId": "ElmAdminController_getRestaurants",
        "summary": "管理端餐馆列表",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "业务管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ElmAdminController_createRestaurant",
        "summary": "管理端创建餐馆",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "业务管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/commerce/restaurants/{id}": {
      "patch": {
        "operationId": "ElmAdminController_updateRestaurant",
        "summary": "管理端更新餐馆",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "业务管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ElmAdminController_deleteRestaurant",
        "summary": "管理端删除餐馆",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "业务管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/commerce/foods": {
      "get": {
        "operationId": "ElmAdminController_getFoods",
        "summary": "管理端食品列表",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "业务管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ElmAdminController_createFood",
        "summary": "管理端创建食品",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "业务管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/commerce/foods/{id}": {
      "patch": {
        "operationId": "ElmAdminController_updateFood",
        "summary": "管理端更新食品",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "业务管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ElmAdminController_deleteFood",
        "summary": "管理端删除食品",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "业务管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/commerce/orders": {
      "get": {
        "operationId": "ElmAdminController_getOrders",
        "summary": "管理端真实支付订单列表",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "业务管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/commerce/orders/{orderNo}": {
      "get": {
        "operationId": "ElmAdminController_getOrderDetail",
        "summary": "管理端订单详情",
        "parameters": [
          {
            "name": "orderNo",
            "required": true,
            "in": "path",
            "schema": {
              "example": "ELMDEMO202606020001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin order detail response envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderDetailHttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "业务管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/commerce/orders/{orderNo}/accept": {
      "post": {
        "operationId": "ElmAdminController_acceptOrder",
        "summary": "管理端订单接单",
        "parameters": [
          {
            "name": "orderNo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "业务管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/commerce/orders/{orderNo}/start-preparing": {
      "post": {
        "operationId": "ElmAdminController_startPreparing",
        "summary": "管理端订单开始制作",
        "parameters": [
          {
            "name": "orderNo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "业务管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/admin/commerce/orders/{orderNo}/complete": {
      "post": {
        "operationId": "ElmAdminController_completeOrder",
        "summary": "管理端订单完成",
        "parameters": [
          {
            "name": "orderNo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "业务管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/commerce/orders/{orderNo}/refund/approve": {
      "post": {
        "operationId": "ElmAdminController_approveRefund",
        "summary": "管理端同意退款",
        "parameters": [
          {
            "name": "orderNo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRefundDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "业务管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/commerce/orders/{orderNo}/refund/reject": {
      "post": {
        "operationId": "ElmAdminController_rejectRefund",
        "summary": "管理端驳回退款",
        "parameters": [
          {
            "name": "orderNo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRefundDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "业务管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/orders": {
      "get": {
        "operationId": "PaymentController_listOrders",
        "summary": "用户支付订单列表",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "支付"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Vue3 Elm API",
    "description": "Vue3 + Elm 项目后端 API 文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AdminOrderActionLogDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "orderNo": {
            "type": "string",
            "example": "ELMDEMO202606020001"
          },
          "operatorId": {
            "type": "string",
            "example": "1"
          },
          "operatorName": {
            "type": "string",
            "example": "admin"
          },
          "operatorType": {
            "type": "string",
            "enum": [
              "ADMIN",
              "CUSTOMER",
              "SYSTEM"
            ],
            "example": "ADMIN"
          },
          "action": {
            "type": "string",
            "example": "ACCEPT"
          },
          "fromFulfillmentStatus": {
            "type": "string",
            "enum": [
              "PENDING_PAYMENT",
              "AWAITING_ACCEPTANCE",
              "ACCEPTED",
              "PREPARING",
              "DELIVERING",
              "COMPLETED",
              "CANCELED"
            ],
            "nullable": true
          },
          "toFulfillmentStatus": {
            "type": "string",
            "enum": [
              "PENDING_PAYMENT",
              "AWAITING_ACCEPTANCE",
              "ACCEPTED",
              "PREPARING",
              "DELIVERING",
              "COMPLETED",
              "CANCELED"
            ],
            "nullable": true
          },
          "fromRefundStatus": {
            "type": "string",
            "enum": [
              "NONE",
              "REQUESTED",
              "APPROVED",
              "REJECTED"
            ],
            "nullable": true
          },
          "toRefundStatus": {
            "type": "string",
            "enum": [
              "NONE",
              "REQUESTED",
              "APPROVED",
              "REJECTED"
            ],
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "remark": {
            "type": "string",
            "nullable": true
          },
          "requestId": {
            "type": "string",
            "example": "req-admin-1",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-06-02T10:00:00.000Z",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "orderNo",
          "operatorId",
          "operatorName",
          "operatorType",
          "action",
          "fromFulfillmentStatus",
          "toFulfillmentStatus",
          "fromRefundStatus",
          "toRefundStatus",
          "reason",
          "remark",
          "requestId",
          "createdAt"
        ]
      },
      "AdminOrderDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "orderNo": {
            "type": "string",
            "example": "ELMDEMO202606020001"
          },
          "userId": {
            "type": "string",
            "example": "42"
          },
          "shopId": {
            "type": "string",
            "example": "1",
            "nullable": true
          },
          "shopName": {
            "type": "string",
            "example": "Demo Shop"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PAID",
              "CLOSED"
            ],
            "example": "PAID"
          },
          "tradeStatus": {
            "type": "string",
            "example": "TRADE_SUCCESS"
          },
          "fulfillmentStatus": {
            "type": "string",
            "enum": [
              "PENDING_PAYMENT",
              "AWAITING_ACCEPTANCE",
              "ACCEPTED",
              "PREPARING",
              "DELIVERING",
              "COMPLETED",
              "CANCELED"
            ],
            "example": "AWAITING_ACCEPTANCE"
          },
          "refundStatus": {
            "type": "string",
            "enum": [
              "NONE",
              "REQUESTED",
              "APPROVED",
              "REJECTED"
            ],
            "example": "NONE"
          },
          "refundBaseFulfillmentStatus": {
            "type": "string",
            "enum": [
              "PENDING_PAYMENT",
              "AWAITING_ACCEPTANCE",
              "ACCEPTED",
              "PREPARING",
              "DELIVERING",
              "COMPLETED",
              "CANCELED"
            ],
            "nullable": true
          },
          "refundReason": {
            "type": "string",
            "nullable": true
          },
          "refundRejectReason": {
            "type": "string",
            "nullable": true
          },
          "tradeNo": {
            "type": "string",
            "nullable": true
          },
          "payableAmount": {
            "type": "number",
            "example": 29
          },
          "goodsAmount": {
            "type": "number",
            "example": 24
          },
          "deliveryFee": {
            "type": "number",
            "example": 5
          },
          "cartItems": {
            "type": "array",
            "items": {
              "additionalProperties": true,
              "type": "object"
            }
          },
          "totalQty": {
            "type": "number",
            "example": 2
          },
          "paidAt": {
            "type": "string",
            "nullable": true
          },
          "acceptedAt": {
            "type": "string",
            "nullable": true
          },
          "preparingAt": {
            "type": "string",
            "nullable": true
          },
          "deliveringAt": {
            "type": "string",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "nullable": true
          },
          "canceledAt": {
            "type": "string",
            "nullable": true
          },
          "refundRequestedAt": {
            "type": "string",
            "nullable": true
          },
          "refundedAt": {
            "type": "string",
            "nullable": true
          },
          "refundRejectedAt": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-06-02T09:55:00.000Z",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-06-02T10:00:00.000Z",
            "format": "date-time"
          },
          "availableActions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "ACCEPT",
                "START_PREPARING",
                "START_DELIVERY",
                "COMPLETE",
                "APPROVE_REFUND",
                "REJECT_REFUND"
              ]
            }
          },
          "customerAvailableActions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "REQUEST_REFUND"
              ]
            }
          },
          "actionLogs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminOrderActionLogDto"
            }
          }
        },
        "required": [
          "id",
          "orderNo",
          "userId",
          "shopId",
          "shopName",
          "status",
          "tradeStatus",
          "fulfillmentStatus",
          "refundStatus",
          "refundBaseFulfillmentStatus",
          "refundReason",
          "refundRejectReason",
          "tradeNo",
          "payableAmount",
          "goodsAmount",
          "deliveryFee",
          "cartItems",
          "totalQty",
          "paidAt",
          "acceptedAt",
          "preparingAt",
          "deliveringAt",
          "completedAt",
          "canceledAt",
          "refundRequestedAt",
          "refundedAt",
          "refundRejectedAt",
          "createdAt",
          "updatedAt",
          "availableActions",
          "customerAvailableActions",
          "actionLogs"
        ]
      },
      "AdminOrderDetailHttpResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/AdminOrderDetailDto"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-06-08T00:00:00.000Z"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "timestamp"
        ]
      },
      "ApproveRefundDto": {
        "type": "object",
        "properties": {}
      },
      "RejectRefundDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}